feat(spotlight): add predictive parser and UI module
