+++ src/ui/mod.rs
@@ -1,2 +1,3 @@
 pub mod beamx;
 pub mod shortcuts;
+pub mod components;
